Reduce Skipped Tests and Test organization
