+++ Dapper.Tests.Database/TestSuitesOracle.cs
@@ -9,7 +9,7 @@ using Dapper.Database.Adapters;
 using Dapper.Database.Extensions;
 using Oracle.ManagedDataAccess.Client;
 using Xunit;
-
+using FactAttribute = Xunit.SkippableFactAttribute;
 using OracleConnection = Dapper.Tests.Database.OracleClient.OracleConnection;
 
 namespace Dapper.Tests.Database
@@ -104,6 +104,33 @@ namespace Dapper.Tests.Database
                 _skip = true;
             }
         }
+
+        #region Oracle Specific Tests
+
+        [Fact]
+        [Trait("Category", "Insert")]
+        public void InsertSequenceComputed()
+        {
+            using (var db = GetSqlDatabase())
+            {
+                var p = new PersonIdentitySequence { FirstName = "Person", LastName = "Identity" };
+                Assert.True(db.Insert(p));
+
+                Assert.True(p.IdentityId > 0);
+                if (p.FullName != null)
+                {
+                    Assert.Equal("Person Identity", p.FullName);
+                }
+
+                var gp = db.Get<PersonIdentitySequence>(p.IdentityId);
+
+                Assert.Equal(p.IdentityId, gp.IdentityId);
+                Assert.Equal(p.FirstName, gp.FirstName);
+                Assert.Equal(p.LastName, gp.LastName);
+                Assert.Equal("Person Identity", gp.FullName);
+            }
+        }
+        #endregion
     }
 }
 #endif
