+++ Dapper.Tests.Database/TestSuitesSqlServer.cs
@@ -1,10 +1,12 @@
-﻿﻿using System;
-using System.Data.SqlClient;
+﻿using System.Data.SqlClient;
 using System.IO;
 using Dapper.Database;
 using Dapper.Database.Adapters;
 using Dapper.Database.Extensions;
 using Xunit;
+using System.Linq;
+using FactAttribute = Xunit.SkippableFactAttribute;
+using System.Threading.Tasks;
 
 namespace Dapper.Tests.Database
 {
@@ -69,5 +71,89 @@ namespace Dapper.Tests.Database
                     throw;
             }
         }
+
+        #region Sql Server Tests Only
+        [Fact]
+        [Trait("Category", "GetMultiple")]
+        public void GetMultiple()
+        {
+            using (var db = GetSqlDatabase())
+            {
+                using (var trans = db.GetTransaction())
+                {
+                    var dt = db.GetMultiple(@"
+                    select * from Product where Color = 'Black';
+                    select * from ProductCategory where productcategoryid = '21';");
+                    Assert.Equal(89, dt.Read(typeof(Product)).Count());
+
+                    var pc = (ProductCategory)dt.ReadSingle(typeof(ProductCategory));
+                    ValidateProductCategory21(pc);
+                    trans.Complete();
+                }
+            }
+        }
+
+        [Fact]
+        [Trait("Category", "GetMultiple")]
+        public void GetMultipleWithParameter()
+        {
+            using (var db = GetSqlDatabase())
+            {
+                using (var trans = db.GetTransaction())
+                {
+                    var dt = db.GetMultiple($@"
+                    select * from Product where Color = {P}Color;
+                    select * from ProductCategory where productcategoryid = {P}ProductCategoryId;",
+                        new { Color = "Black", ProductCategoryId = 21 });
+                    Assert.Equal(89, dt.Read(typeof(Product)).Count());
+
+                    var pc = (ProductCategory)dt.ReadSingle(typeof(ProductCategory));
+                    ValidateProductCategory21(pc);
+                    trans.Complete();
+                }
+            }
+        }
+
+        [Fact]
+        [Trait("Category", "GetMultipleAsync")]
+        public async Task GetMultipleAsync()
+        {
+            using (var db = GetSqlDatabase())
+            {
+                using (var trans = db.GetTransaction())
+                {
+                    var dt = await db.GetMultipleAsync(@"
+                    select * from Product where Color = 'Black';
+                    select * from ProductCategory where productcategoryid = '21';");
+                    Assert.Equal(89, dt.Read(typeof(Product)).Count());
+
+                    var pc = (ProductCategory)dt.ReadSingle(typeof(ProductCategory));
+                    ValidateProductCategory21(pc);
+                    trans.Complete();
+                }
+            }
+        }
+
+        [Fact]
+        [Trait("Category", "GetMultipleAsync")]
+        public async Task GetMultipleAsyncWithParameter()
+        {
+            using (var db = GetSqlDatabase())
+            {
+                using (var trans = db.GetTransaction())
+                {
+                    var dt = await db.GetMultipleAsync($@"
+                    select * from Product where Color = {P}Color;
+                    select * from ProductCategory where productcategoryid = {P}ProductCategoryId;",
+                        new { Color = "Black", ProductCategoryId = 21 });
+                    Assert.Equal(89, dt.Read(typeof(Product)).Count());
+
+                    var pc = (ProductCategory)dt.ReadSingle(typeof(ProductCategory));
+                    ValidateProductCategory21(pc);
+                    trans.Complete();
+                }
+            }
+        }
+        #endregion
     }
 }
