+++ Dapper.Tests.Database/Tests/TestSuiteExistsAsync.cs
@@ -60,11 +60,7 @@ namespace Dapper.Tests.Database
         {
             using (var db = GetSqlDatabase())
             {
-                if (GetProvider() == Provider.SQLite)
-                {
-                    return;
-                }
-                else if (GetProvider() == Provider.Firebird)
+                if (GetProvider() == Provider.Firebird || GetProvider() == Provider.SQLite)
                 {
                     Assert.True(await db.ExistsAsync<Product>($"where rowguid = {P}GuidId", new { GuidId = "23B5D52B-8C29-4059-B899-75C53B5EE2E6" }));
                     Assert.False(await db.ExistsAsync<Product>($"where rowguid = {P}GuidId", new { GuidId = "1115D52B-8C29-4059-B899-75C53B5EE2E6" }));
