+++ Dapper.Tests.Database/Tests/TestSuiteGetDataTable.cs
@@ -10,13 +10,12 @@ namespace Dapper.Tests.Database
         [Trait("Category", "GetDataTable")]
         public void GetDataTable()
         {
-            if (GetProvider() != Provider.SQLite)
+            Skip.If(GetProvider() == Provider.SQLite, "Sqlite specified method is not supported.");
+
+            using (var db = GetSqlDatabase())
             {
-                using (var db = GetSqlDatabase())
-                {
-                    var dt = db.GetDataTable("select * from Product where Color = 'Black'");
-                    Assert.Equal(89, dt.Rows.Count);
-                }
+                var dt = db.GetDataTable("select * from Product where Color = 'Black'");
+                Assert.Equal(89, dt.Rows.Count);
             }
         }
 
@@ -24,13 +23,12 @@ namespace Dapper.Tests.Database
         [Trait("Category", "GetDataTable")]
         public void GetDataTableWithParameter()
         {
-            if (GetProvider() != Provider.SQLite)
+            Skip.If(GetProvider() == Provider.SQLite, "Sqlite specified method is not supported.");
+
+            using (var db = GetSqlDatabase())
             {
-                using (var db = GetSqlDatabase())
-                {
-                    var dt = db.GetDataTable($"select * from Product where Color = {P}Color", new { @Color = "Black" });
-                    Assert.Equal(89, dt.Rows.Count);
-                }
+                var dt = db.GetDataTable($"select * from Product where Color = {P}Color", new { @Color = "Black" });
+                Assert.Equal(89, dt.Rows.Count);
             }
         }
     }
