+++ Dapper.Tests.Database/Tests/TestSuiteGetList.cs
@@ -109,7 +109,7 @@ namespace Dapper.Tests.Database
             using (var db = GetSqlDatabase())
             {
                 var lst = db.GetList<Product, ProductCategory>(
-                    getListMultiTwoParamQuery, new { Color = "Black" }, "ProductCategoryId");
+                    GetListMultiTwoParamQuery, new { Color = "Black" }, "ProductCategoryId");
                 Assert.Equal(89, lst.Count());
                 var item = lst.Single(p => p.ProductID == 816);
                 ValidateProduct816(item);
@@ -134,7 +134,7 @@ namespace Dapper.Tests.Database
                         pr.ProductCategory = pc;
                         return pr;
                     },
-                   getListMultiTwoParamQuery, new { Color = "Black" }, "ProductCategoryId");
+                   GetListMultiTwoParamQuery, new { Color = "Black" }, "ProductCategoryId");
                 Assert.Equal(89, lst.Count());
                 var item = lst.Single(p => p.ProductID == 816);
                 ValidateProduct816(item);
@@ -154,7 +154,7 @@ namespace Dapper.Tests.Database
             using (var db = GetSqlDatabase())
             {
                 var lst = db.GetList<Product, ProductCategory, ProductModel>(
-                    getListMultiThreeParamQuery, new { Color = "Black" }, "ProductCategoryId,ProductModelId");
+                    GetListMultiThreeParamQuery, new { Color = "Black" }, "ProductCategoryId,ProductModelId");
                 Assert.Equal(89, lst.Count());
                 var item = lst.Single(p => p.ProductID == 816);
                 ValidateProduct816(item);
@@ -181,7 +181,7 @@ namespace Dapper.Tests.Database
                         pr.ProductModel = pm;
                         return pr;
                     },
-                    getListMultiThreeParamQuery, new { Color = "Black" }, "ProductCategoryId,ProductModelId");
+                    GetListMultiThreeParamQuery, new { Color = "Black" }, "ProductCategoryId,ProductModelId");
                 Assert.Equal(89, lst.Count());
                 var item = lst.Single(p => p.ProductID == 816);
                 ValidateProduct816(item);
@@ -192,31 +192,6 @@ namespace Dapper.Tests.Database
                     ValidateProductCategory21(item.ProductCategory);
                     ValidateProductModel45(item.ProductModel);
                 }
-            }
-        }
-
-        private string getListMultiTwoParamQuery
-        {
-            get
-            {
-                return
-                    $@"select  P.*, P.rowguid as GuidId, PC.*
-                    from Product P
-                    join ProductCategory PC on PC.ProductCategoryID = P.ProductCategoryID
-                    where P.Color = {P}Color";
-            }
-        }
-
-        private string getListMultiThreeParamQuery
-        {
-            get
-            {
-                return
-                    $@"select  P.*, P.rowguid as GuidId, PC.*, PM.*
-                    from Product P
-                    join ProductCategory PC on PC.ProductCategoryID = P.ProductCategoryID
-                    join ProductModel PM on PM.ProductModelID = P.ProductModelID
-                    where P.Color = {P}Color";
             }
         }
     }
